+++ puresasl/__init__.py
@@ -1,5 +1,5 @@
-__version__ = '0.4.0'
-__version_info__ = (0, 4, 0)
+__version__ = '0.4.1'
+__version_info__ = (0, 4, 1)
 
 
 class SASLError(Exception):
@@ -15,6 +15,13 @@ class SASLProtocolException(Exception):
     Raised when an error occurs while communicating with the SASL server
     or the client and server fail to agree on negotiated properties such
     as quality of protection.
+    """
+    pass
+
+
+class SASLWarning(Warning):
+    """
+    Emitted in potentially fatal circumstances.
     """
     pass