+++ CHANGES.rst
@@ -1,3 +1,12 @@
+0.4.1
+=====
+March 8th, 2018
+
+* Fix QOP checks in python3 (#19)
+* Improved error handling when the kerberos module is not installed (#20)
+* Fix python3 bug using auth-conf with GSSAPI (#21)
+* Add GSSAPI "extra" with kerberos dependency
+
 0.4.0
 =====
 February 21st, 2017